Enhance PaymentSheetViewModel
